+++ src/assets/index.js
@@ -6,5 +6,9 @@ import lightmode from "./lightmode.png";
 import darkmode from "./darkmode.png";
 
 import hero1 from "./P1232212.jpg";
+import hero2 from "./P1175226.jpg";
+import hero3 from "./P1094314.jpg";
 
-export { logo, instagram, linkedin, unsplash, lightmode, darkmode, hero1 };
+const heroList = [hero1, hero2, hero3];
+
+export { logo, instagram, linkedin, unsplash, lightmode, darkmode, heroList };
