+++ src/components/Contact.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Contact = () => {
+	return <div>Contact</div>;
+};
+
+export default Contact;
