+++ src/components/index.js
@@ -1,8 +1,9 @@
 import Navbar from "./Navbar";
+import Contact from "./Contact";
 import Hero from "./Hero";
 import Gallery from "./Gallery";
 import Footer from "./Footer";
 
 import Spinner from "./Spinner";
 
-export { Navbar, Hero, Gallery, Footer, Spinner };
+export { Navbar, Contact, Hero, Gallery, Footer, Spinner };
