+++ src/gallery/index.js
@@ -1,0 +1,53 @@
+import photo1 from "./P1232212.jpg";
+import photo2 from "./P1049668.jpg";
+import photo3 from "./P1050348.jpg";
+import photo4 from "./P1050420.jpg";
+import photo5 from "./P1050468.jpg";
+import photo6 from "./P1050604.jpg";
+import photo7 from "./P1083411.jpg";
+import photo8 from "./P1083443.jpg";
+import photo9 from "./P1083447.jpg";
+import photo10 from "./P1083782.jpg";
+import photo11 from "./P1083862.jpg";
+import photo12 from "./P1094314.jpg";
+import photo13 from "./P1106223.jpg";
+import photo14 from "./P1106227.jpg";
+import photo15 from "./P1128772.jpg";
+import photo16 from "./P1128832.jpg";
+import photo17 from "./P1128924.jpg";
+import photo18 from "./P1139979.jpg";
+import photo19 from "./P1140028.jpg";
+import photo20 from "./P1151587.jpg";
+import photo21 from "./P1151740.jpg";
+import photo22 from "./P1175226.jpg";
+import photo23 from "./P1208181.jpg";
+import photo24 from "./P1208191.jpg";
+
+const gallery = [
+	photo1,
+	photo2,
+	photo3,
+	photo4,
+	photo5,
+	photo6,
+	photo7,
+	photo8,
+	photo9,
+	photo10,
+	photo11,
+	photo12,
+	photo13,
+	photo14,
+	photo15,
+	photo16,
+	photo17,
+	photo18,
+	photo19,
+	photo20,
+	photo21,
+	photo22,
+	photo23,
+	photo24,
+];
+
+export { gallery };
